+++ qnum.h
@@ -1,8 +1,8 @@
 /* See LICENSE file for copyright and license details. */
 
 typedef struct {
-	long num;
-	long den;
+	long long num;
+	long long den;
 } qnum;
 
 void qnum_reduce(qnum *n);
